requirements.txt
httpx>=0.27
pypdf>=4.0
python-dotenv>=1.0
google-auth>=2.29
google-auth-oauthlib>=1.2
google-api-python-client>=2.126
